osx fixes for go 1.11
This is a 3-in-1 story. Please have a seat.

1. As Go 1.11 release notes say,

> Go 1.11 now requires <...> macOS 10.10 Yosemite or later

So, change the OSX_VERSION_MIN in osxcross to 10.10.

2. MACOSX SDK tarball that was used before comes without libc++
headers, which are now required by at least crypto/x509 pkg
when compiling for Darwin. These include files are part of xcode,
which can only be downloaded by someone having Apple Developer ID,
which I don't have so this path is not available for me.

An alternative is to reuse the SDK tarball from the
https://github.com/multiarch/crossbuild project. Using it
fixes the following error:

> # crypto/x509
> osxcross: error: cannot find libc++ headers

3. The new tarball, although, misses the libtool package (apparently
somehow injected into the tarball that was used earlier) needed to
compile miekg/pkcs11 used by cli, which leads to errors like:

> # github.com/docker/cli/vendor/github.com/miekg/pkcs11
> vendor/github.com/miekg/pkcs11/pkcs11.go:28:10: fatal error: 'ltdl.h' file not found
> #include <ltdl.h>
>          ^
> 1 error generated.

To fix this one, libtool for osx need to be installed. Building it from
source (with osxcross) requires some other dependencies. Using brew is not
possible since it is Linux. Solution: download a binary package from brew
repo, cherry pick the needed stuff (lib and include).

Tested to fix docker-cli compilation issues on OSX.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>