Update to Go 1.13.8
This updates the image for Go 1.13. Note that the Apple SDK that
is hosted on s3.mobyproject.org is an older version, which does
not allow specifying the "minimum macOS version" to be 1.11 (the
minimum version supported by Go 1.13).

TODO comments were added to the Dockerfile to document this, and
to update the version once an updated SDK is available.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>